Testing error

+++ tests/smoke/mautic.sh
@@ -1,5 +1,12 @@
 echo "[Smoke] - Mautic"
 
 # docker-compose up -d mautic
-curl --connect-timeout 30 --fail -I http://localhost:8082
+sleep 30
+
+HTTP_STATUS=$(curl -o /dev/null -s -w "%{http_code}\n" http://localhost:8082)
+if [ $HTTP_STATUS -ne 302  ];
+  then
+    exit 1
+fi
+
 docker-compose down
